fixing login and logout

+++ index.php
@@ -19,12 +19,13 @@ session_start();
     <div class="container">
 
         <div class="row">
-            <div class="col-2 text-right">
+
+            <div class="text-end">
                 <?php
                 if (isset($_SESSION['email'])) {
                     // User is logged in, display the logout button.
                     echo "Welcome, " . $_SESSION['username'] . "!<br>";
-                    echo "<a href='logout.php'>Logout</a>";
+                    echo "<a href='../logout.php'>Logout</a>";
                 } else {
                     // User is not logged in, display registration and login forms.
                     include('forms.php');
@@ -32,48 +33,51 @@ session_start();
                 ?>
             </div>
 
-            <div class="col-10">
+            <div class="col-12">
 
 
                 <?php
-                $serialziedData = file_get_contents('./data/db.txt');
-                $students = unserialize($serialziedData);
+                if (isset($_SESSION['email'])) { ?>
 
-                ?>
-
-                <table class="table table-bordred">
-                    <tr>
-                        <th>Name</th>
-                        <th>Roll</th>
-
-                        <th width="25%">Action</th>
-
-                    </tr>
                     <?php
-                    foreach ($students as $student) {
-                        ?>
+                    $serialziedData = file_get_contents('../data/db.txt');
+                    $students = unserialize($serialziedData);
+                    ?>
+                    <table class="table table-bordred">
                         <tr>
-                            <td>
-                                <?php printf('%s %s', $student['fname'], $student['lname']); ?>
-                            </td>
-                            <td>
-                                <?php printf('%s', $student['roll']); ?>
-                            </td>
+                            <th>Name</th>
+                            <th>Roll</th>
 
-                            <td>
-                                <?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a> | <a class="delete" href="/crud/index.php?task=delete&id=%s">Delete</a>', $student['id'], $student['id']); ?>
-                            </td>
-
-                            <td>
-                                <?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a>', $student['id']); ?>
-                            </td>
+                            <th width="25%">Action</th>
 
                         </tr>
                         <?php
-                    }
-                    ?>
+                        foreach ($students as $student) {
+                            ?>
+                            <tr>
+                                <td>
+                                    <?php printf('%s %s', $student['fname'], $student['lname']); ?>
+                                </td>
+                                <td>
+                                    <?php printf('%s', $student['roll']); ?>
+                                </td>
 
-                </table>
+                                <td>
+                                    <?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a> | <a class="delete" href="/crud/index.php?task=delete&id=%s">Delete</a>', $student['id'], $student['id']); ?>
+                                </td>
+
+                                <td>
+                                    <?php printf('<a href="/crud/index.php?task=edit&id=%s">Edit</a>', $student['id']); ?>
+                                </td>
+
+                            </tr>
+                            <?php
+                        }
+                        ?>
+
+                    </table>
+                <?php } ?>
+
             </div>
         </div>
